Productivo/requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery